chore(cloudflare): optimize Cache usage; strict TS source
