+++ packages/adapter-cloudflare/files/worker.ts
@@ -1,0 +1,74 @@
+import { Server } from 'SERVER';
+import { manifest, prerendered } from 'MANIFEST';
+import * as Cache from 'worktop/cfw.cache';
+
+import type { Module, Bindings } from 'worktop/cfw';
+import type { Durable } from 'worktop/cfw.durable';
+
+interface Environment extends Bindings {
+	ASSETS: Durable.Object;
+}
+
+const server = new Server(manifest);
+
+const prefix = `/${manifest.appDir}/`;
+
+const worker: Module.Worker<Environment> = {
+	async fetch(req, env, context) {
+		try {
+			let res = await Cache.lookup(req);
+			if (res) return res;
+
+			let { pathname } = new URL(req.url);
+
+			// static assets
+			if (pathname.startsWith(prefix)) {
+				res = await env.ASSETS.fetch(req);
+
+				res = new Response(res.body, {
+					headers: {
+						// include original cache headers, minus cache-control which
+						// is overridden, and etag which is no longer useful
+						'cache-control': 'public, immutable, max-age=31536000',
+						'content-type': res.headers.get('content-type'),
+						'x-robots-tag': 'noindex'
+					}
+				});
+			} else {
+				// prerendered pages and index.html files
+				pathname = pathname.replace(/\/$/, '') || '/';
+
+				let file = pathname.substring(1);
+
+				try {
+					file = decodeURIComponent(file);
+				} catch (err) {
+					// ignore
+				}
+
+				if (
+					manifest.assets.has(file) ||
+					manifest.assets.has(file + '/index.html') ||
+					prerendered.has(pathname)
+				) {
+					res = await env.ASSETS.fetch(req);
+				} else {
+					// dynamically-generated pages
+					res = await server.respond(req, {
+						platform: { env, context },
+						getClientAddress() {
+							return req.headers.get('cf-connecting-ip');
+						}
+					});
+				}
+			}
+
+			// Writes to Cache only if allowed
+			return Cache.save(req, res, context);
+		} catch (e) {
+			return new Response('Error rendering route: ' + (e.message || e.toString()), { status: 500 });
+		}
+	}
+};
+
+export default worker;
