+++ packages/adapter-cloudflare/tsconfig.json
@@ -11,5 +11,5 @@
 			"@sveltejs/kit": ["../kit/types/index"]
 		}
 	},
-	"include": ["**/*.js", "ambient.d.ts"]
+	"include": ["**/*.js", "ambient.d.ts", "files/worker.ts"]
 }
